# apps/backend/docs/openapi/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  modules_compliance_api.groSettings:
    properties:
      gro_email:
        type: string
      gro_name:
        type: string
      gro_phone:
        type: string
      is_significant_data_fiduciary:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: security@arc-hawk.io
    name: ARC-Hawk Security Team
  description: Enterprise PII Discovery, Classification, and Lineage Tracking Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://arc-hawk.io/terms
  title: ARC-HAWK-DD API
  version: 3.0.0
paths:
...
  /api/v1/compliance/audit-trail:
    get:
      consumes:
      - application/json
      description: Returns paginated audit ledger rows for the tenant
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Filter by event type (repeatable)
        in: query
        name: event_type
        type: string
      - description: Start time in RFC3339 format
        in: query
        name: from
        type: string
      - description: End time in RFC3339 format
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get compliance audit trail
      tags:
      - compliance
...
  /api/v1/compliance/evidence-package:
    post:
      consumes:
      - application/json
      description: Generates a ZIP containing all DPDP evidence and streams it to
        the client
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Generate DPDP evidence package
      tags:
      - compliance
  /api/v1/compliance/gro-settings:
    get:
      consumes:
      - application/json
      description: Returns the current tenant's Grievance Redressal Officer settings
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get GRO settings
      tags:
      - compliance
    put:
      consumes:
      - application/json
      description: Updates the Grievance Redressal Officer configuration for the tenant
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: GRO settings payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/modules_compliance_api.groSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update GRO settings
      tags:
      - compliance
...
  /discovery/inventory:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List full asset inventory with source mapping
      tags:
      - discovery
  /discovery/inventory/{assetId}:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Asset UUID
        in: path
        name: assetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get per-asset inventory detail
      tags:
      - discovery
  /discovery/overview:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get discovery dashboard overview
      tags:
      - discovery
...
securityDefinitions:
  APIKeyAuth:
    description: Tenant API key for service-to-service authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"